@@ -1,416 +1,258 @@
-import React from "react";
-function about() {
-  return (
-    //About Section
-    <section id="about" classname="about-section text-center">
-      <div classname="container">
-        <div classname="row">
-          <div classname="col-lg-8 mx-auto">
-            <h2 classname="text-white mb-4">
-              Working throughout the year to build relationships between
-              students and alumni.
-            </h2>
-            <hr classname="accessory" />
-            <p classname="text-white-50">
-              The Union College Student Alumni Association, The Garnet Society,
-              is one of the largest student run organizations on campus today.
-              TGS is committed to connecting students to Union College alumni
-              and the larger community through a variety of programs and events
-              throughout the year. TGS acts as ambassadors at Union College
-              functions and help facilitate student philanthropy efforts.{" "}
-            </p>
-            <img
-              src="https://i.postimg.cc/d1PsJ8kR/union.png"
-              classname="col-lg-6 img-fluid unionicon"
-              alt=""
-            />{" "}
-            <img
-              src="https://i.postimg.cc/Kc7zvs51/tgscrestinverted.png"
-              classname="col-lg-6 img-fluid cresticon"
-              alt=""
-            />
-            <p classname="text-white-50">
-              TGS plays a huge role during major events hosted on campus by The
-              Office of Alumni & Parent Engagement & other divisions within
-              College Relations. Members of TGS also benefit from the great
-              leadership skills they learn, and they receive a behind the scenes
-              look at many of our top events. Members have the opportunity to
-              interact with an array of alumni and gain great networking skills
-              hosting alumni events as student ambassadors.{" "}
-            </p>
-            <h3 classname="text-white mb-4">Meet our members</h3>
-            <hr classname="accessory" width="50%" />
-            <br />
-          </div>
-        </div>
+import React from 'react';
+function about(){
+return(
+    //  About Section
+  <section id="about" className="about-section text-center">
+    <div className="container">
+      <div className="row">
+        <div className="col-lg-8 mx-auto">
+          <h2 className="text-white mb-4">Working throughout the year to build relationships between students and alumni.</h2>
+          <hr className="accessory"/>
+          <p className="text-white-50">The Union College Student Alumni Association, The Garnet Society, is one of the largest student run organizations on campus today. TGS is committed to connecting students to Union College alumni and the larger community through a variety of
+            programs and events throughout the year. TGS acts as ambassadors at Union College functions and help facilitate student philanthropy efforts. </p>
+          <img src='https://i.postimg.cc/d1PsJ8kR/union.png' className="col-lg-6 img-fluid unionicon" alt=""/> <img src="https://i.postimg.cc/Kc7zvs51/tgscrestinverted.png" className="col-lg-6 img-fluid cresticon" alt=""/>
 
-        <br></br>
-      </div>
-      //Headshots
-      <div classname="container-fluid">
-        //row 1
-        <div classname="row headshots">
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                classname="headshotphoto"
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: VP of Recruitment
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: Junior{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: Economics & Political Science
-              </p>
-            </figure>
-          </div>
+          <p className="text-white-50">TGS plays a huge role during major events hosted on campus by The Office of Alumni & Parent Engagement & other divisions within College Relations. Members of TGS also benefit from the great leadership skills they learn, and they receive a behind
+            the scenes look at many of our top events. Members have the opportunity to interact with an array of alumni and gain great networking skills hosting alumni events as student ambassadors. </p>
 
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                classname="headshotphoto"
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A
-              </p>
-            </figure>
-          </div>
+          <h3 className="text-white mb-4">Meet our members</h3>
 
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                classname="headshotphoto"
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
+          <hr className="accessory" width="50%"/>
+          <br/>
 
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                classname="headshotphoto"
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
-        </div>
-        // Row 2
-        <div classname="row">
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
-        </div>
-        //Row 3
-        <div classname="row">
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
-        </div>
-        //Row 4
-        <div classname="row">
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Female</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
-
-          <div classname="col-sm-3">
-            <figure classname="one">
-              <img
-                src="https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg"
-                alt=""
-              />
-              <p classname="text-white headshottext">Example Male</p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Position</b>: N/A
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Year</b>: N/A{" "}
-              </p>
-              <p classname="text-white headshotgrade">
-                {" "}
-                <b>Major</b>: N/A{" "}
-              </p>
-            </figure>
-          </div>
         </div>
       </div>
-      //Container div
-    </section>
-  );
+
+  <br/><br/>
+
+    </div>
+    {/* // Headshots */}
+    <div className="container-fluid">
+
+      {/* //row 1 */}
+      <div className="row headshots">
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' className="headshotphoto" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: VP of Recruitment</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: Junior </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: Economics & Political Science</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' className="headshotphoto" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' className="headshotphoto" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' className="headshotphoto" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+      </div>
+
+      {/* // Row 2 */}
+      <div className="row">
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+      </div>
+
+      {/* //Row 3  */}
+      <div className="row">
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+      </div>
+
+      {/* //Row 4 */}
+      <div className="row">
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A</p>
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/tCfL09VW/femaleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Female</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+        <div className="col-sm-3">
+
+          <figure className="one">
+            <img src='https://i.postimg.cc/CxCTbWz8/maleheadshot.jpg' alt="" />
+            <p className="text-white headshottext">Example Male</p>
+            <p className="text-white headshotgrade"> <b>Position</b>: N/A</p>
+            <p className="text-white headshotgrade"> <b>Year</b>: N/A </p>
+            <p className="text-white headshotgrade"> <b>Major</b>: N/A </p>
+
+          </figure>
+
+        </div>
+
+      </div>
+
+   </div> 
+   <br/><br/>
+  </section>
+
+);
 }
 export default about;
