@@ -1,7 +1,9 @@
-import React from "react";
-function foot() {
-  <footer classname="bg-black small text-center text-white-50">
-    <div classname="container">Copyright &copy; Tucker Massad 2019</div>
-  </footer>;
+import React from 'react';
+function foot(){
+<footer className="bg-black small text-center text-white-50">
+    <div className="container">
+      Copyright &copy; Tucker Massad 2019
+    </div>
+  </footer>
 }
 export default foot;
