@@ -1,17 +1,15 @@
-import React from "react";
-function header() {
-  <header classname="masthead">
-    <div classname="container d-flex h-100 align-items-center">
-      <div classname="mx-auto text-center">
-        <h1 classname="mx-auto my-0 text-uppercase">The Garnet Society</h1>
-        <h2 classname="text-white-50 mx-auto mt-2 mb-5">
-          The Union College Student Alumni Association
-        </h2>
-        <a href="#about" classname="btn btn-primary js-scroll-trigger">
-          Learn More
-        </a>
+import React from 'react';
+function header(){
+  return(
+<header className="masthead">
+    <div className="container d-flex h-100 align-items-center">
+      <div className="mx-auto text-center">
+        <h1 className="mx-auto my-0 text-uppercase">The Garnet Society</h1>
+        <h2 className="text-white-50 mx-auto mt-2 mb-5">The Union College Student Alumni Association</h2>
+        <a href="#about" className="btn btn-primary js-scroll-trigger">Learn More</a>
       </div>
     </div>
-  </header>;
+  </header>
+  );
 }
 export default header;
